Update README.md
The sediment transport caused by the flow of an extreme flood, such as the propagation of a dam failure, is being studied to simulate the hydrodynamic behaviour of the solid particles contained in the valley during the flood passage. With this intention, I have developed a numerical model which is based on the resolution of the one-dimensional Saint Venant–Exner equations by the implicit finite difference scheme.